@@ -3,18 +3,18 @@
     "version": "5.2.6",
     "private": true,
     "dependencies": {
-        "@iobroker/adapter-react-v5": "^2.1.9",
-        "@mui/material": "^5.8.0",
-        "@mui/icons-material": "^5.8.0",
-        "@mui/styles": "^5.8.0",
+        "@iobroker/adapter-react-v5": "^3.1.33",
+        "@mui/material": "^5.10.2",
+        "@mui/icons-material": "^5.10.2",
+        "@mui/styles": "^5.10.2",
         "babel-eslint": "^10.1.0",
-        "@sentry/browser": "^6.19.7",
-        "@sentry/integrations": "^6.19.7",
-        "copy-to-clipboard": "^3.3.1",
-        "react": "^18.1.0",
-        "react-dom": "^18.1.0",
-        "react-dropzone": "^14.2.1",
-        "react-icons": "^4.3.1",
+        "@sentry/browser": "^7.11.1",
+        "@sentry/integrations": "^7.11.1",
+        "copy-to-clipboard": "^3.3.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-dropzone": "^14.2.2",
+        "react-icons": "^4.4.0",
         "react-scripts": "^5.0.1"
     },
     "scripts": {
